+++ domestic/helpers.py
@@ -1,13 +1,13 @@
 from sso import helpers as sso_helpers
-from core.models import DetailPage, ListPage
+from core.models import DetailPage, CuratedListPage
 
 
 def check_route(route_type, context, user_profile):
     # Check if this route is valid to show given user context
     if route_type == 'learn':
         total_read = 0
-        for list_page in context.get('list_pages'):
-            total_read += list_page.get('read_count')
+        for module_page in context.get('module_pages'):
+            total_read += module_page.get('read_count')
         return total_read == 0
 
     elif route_type == 'target':
@@ -55,17 +55,17 @@ def get_read_progress(user, context={}):
         completed.add(lesson.get('lesson'))
     page_map = {}
     for detail_page in DetailPage.objects.live():
-        list_page = get_ancestor(detail_page, ListPage)
-        if list_page:
-            page_map[list_page.id] = page_map.get(
-                list_page.id) or {'total_pages': 0, 'read_count': 0, 'page': list_page}
-            page_map[list_page.id]['total_pages'] += 1
+        module_page = get_ancestor(detail_page, CuratedListPage)
+        if module_page:
+            page_map[module_page.id] = page_map.get(
+                module_page.id) or {'total_pages': 0, 'read_count': 0, 'page': module_page}
+            page_map[module_page.id]['total_pages'] += 1
             if detail_page.id in completed:
-                page_map[list_page.id]['read_count'] += 1
+                page_map[module_page.id]['read_count'] += 1
                 lessons_in_progress = True
-    list_pages = list(page_map.values())
-    list_pages.sort(key=lesson_comparator, reverse=True)
+    module_pages = list(page_map.values())
+    module_pages.sort(key=lesson_comparator, reverse=True)
     return {
-        'list_pages': list_pages,
+        'module_pages': module_pages,
         'lessons_in_progress': lessons_in_progress
     }